Code refactoring for adapters
