+++ app/src/main/java/com/wentongwang/mysports/constant/Constant.java
@@ -8,14 +8,14 @@ public class Constant {
     //*********************** local host *******************************/
     //this is for test, in fact, should save in SP
 //    public static final String HOST = "http://35.156.49.76:8080/sports";
-    public static final String HOST = "http://35.156.49.76:8080/sports";
+    public static final String HOST = "http://192.168.1.35:8080/sports";
 
     //*********************** servlet path *******************************//
     public static final String LOGIN_PATH = "/user/login";
     public static final String SIGN_UP_PATH = "/user/register";
     public static final String GET_NEWS_PATH = "/articles/getarticles";
     public static final String GET_AGENDA_PATH="/agenda/getagenda";
-
+    public static final String GET_EVENT_PATH="/home/gethomesportevent";
 
     //测试提交1
     //测试提交2
