+++ app/src/main/java/com/wentongwang/mysports/views/activity/home/HomeFragmentPagerAdapter.java
@@ -1,0 +1,30 @@
+package com.wentongwang.mysports.views.activity.home;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+import java.util.List;
+
+/**
+ * Created by qifan on 2017/3/25.
+ */
+
+public class HomeFragmentPagerAdapter extends FragmentPagerAdapter {
+    List<Fragment> list;
+
+    public HomeFragmentPagerAdapter(FragmentManager fm, List<Fragment> list) {
+        super(fm);
+        this.list=list;
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        return list.get(position);
+    }
+
+    @Override
+    public int getCount() {
+        return list.size();
+    }
+}
