+++ app/src/main/java/com/wentongwang/mysports/views/fragment/agenda/AgendaFragment.java
@@ -32,7 +32,7 @@ public class AgendaFragment extends BaseFragment implements AgendaView {
     @BindView(R.id.agenda_events_list)
     protected ListView listView;
 
-    private MyAdapter adapter;
+    private AgendaEventsAdapter adapter;
 
     private AgendaPresenter mPresenter = new AgendaPresenter(this);
 
@@ -44,7 +44,7 @@ public class AgendaFragment extends BaseFragment implements AgendaView {
     @Override
     public void initDatas() {
         mPresenter.init(getActivity());
-        adapter = new MyAdapter();
+        adapter = new AgendaEventsAdapter();
         listView.setAdapter(adapter);
         mPresenter.getAgenda();
     }
@@ -80,74 +80,6 @@ public class AgendaFragment extends BaseFragment implements AgendaView {
     public void setAgendaList(List<AgendaEvents> list) {
         adapter.setItemList(list);
         adapter.notifyDataSetChanged();
-    }
-
-    class MyAdapter extends BaseAdapter {
-
-        private List<AgendaEvents> itemList = new ArrayList<>();
-
-        public void setItemList(List<AgendaEvents> list) {
-            this.itemList = list;
-        }
-
-        @Override
-        public int getCount() {
-            return itemList.size();
-        }
-
-        @Override
-        public Object getItem(int position) {
-            return itemList.get(position);
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            ViewHolder viewHolder = null;
-            if (convertView == null) {
-                convertView = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_agenda_event_list, null);
-                viewHolder = new ViewHolder(convertView);
-
-                convertView.setTag(viewHolder);
-            } else {
-                viewHolder = (ViewHolder) convertView.getTag();
-            }
-            AgendaEvents agendaEvents = itemList.get(position);
-            viewHolder.event_place.setText(agendaEvents.getUser_events_place());
-            viewHolder.event_time.setText(agendaEvents.getUser_events_time());
-            viewHolder.event_type.setText(agendaEvents.getUser_events_name());
-            viewHolder.event_progress.setProgress(Integer.parseInt(agendaEvents.getEvents_Progress()));
-            viewHolder.user_head.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    mPresenter.showPopupWindow();
-                }
-            });
-
-            return convertView;
-        }
-
-        class ViewHolder {
-            @BindView(R.id.user_head)
-            CircleImageView user_head;
-            @BindView(R.id.tv_event_type)
-            TextView event_type;
-            @BindView(R.id.tv_event_time)
-            TextView event_time;
-            @BindView(R.id.tv_event_place)
-            TextView event_place;
-            @BindView(R.id.event_progress_bar)
-            MyProgressBarHorizontal event_progress;
-
-            public ViewHolder(View v) {
-                ButterKnife.bind(this, v);
-
-            }
-        }
     }
