+++ app/src/main/java/com/wentongwang/mysports/views/fragment/friendrequest/FriendRequestFragment.java
@@ -20,7 +20,7 @@ public class FriendRequestFragment extends BaseFragment {
     @BindView(R.id.grid_view)
     protected NoScrollGridView gridView;
 
-    private MyGridViewAdapter adapter;
+    private GridViewAdapter adapter;
     @Override
     public int getLayoutId() {
         return R.layout.friend_request_fragment_layout;
@@ -28,7 +28,7 @@ public class FriendRequestFragment extends BaseFragment {
 
     @Override
     public void initDatas() {
-        adapter = new MyGridViewAdapter();
+        adapter = new GridViewAdapter();
         gridView.setAdapter(adapter);
     }
 
@@ -37,46 +37,5 @@ public class FriendRequestFragment extends BaseFragment {
 
     }
 
-    class MyGridViewAdapter extends BaseAdapter {
-        @Override
-        public int getCount() {
-            return 14;
-        }
-
-        @Override
-        public Object getItem(int position) {
-            return null;
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            ViewHolder viewHolder = null;
-            if (convertView == null) {
-                convertView = LayoutInflater.from(getActivity()).inflate(R.layout.item_friend_request_gridview, null);
-                viewHolder = new ViewHolder(convertView);
-
-                convertView.setTag(viewHolder);
-            } else {
-                viewHolder = (ViewHolder) convertView.getTag();
-            }
-
-
-            return convertView;
-        }
-
-        class ViewHolder{
-
-
-            public ViewHolder(View view){
-
-            }
-
-        }
-    }
 
 }
